pyproject.toml
[tool.poetry]
name = "drf-poc"
version = "0.1.0"
description = "This is a project that aims to test some usage concepts, techniques and technologies."
authors = ["atmosmps <atmos.mps@gmail.com>"]
license = "GNU General Public License"

[tool.poetry.dependencies]
python = "^3.7"
asgiref = "^3.4"
attrs = "^21.2"
dj-database-url = "^0.5.0"
dj-static = "^0.0.6"
Django = "^3.2"
django-filter = "^21.1"
djangorestframework = "^3.12"
djangorestframework-simplejwt = "^5.0"
drf-spectacular = "^0.20"
gunicorn = "^20.1"
importlib-metadata = "^4.2"
inflection = "^0.5"
install = "^1.3"
jsonschema = "^4.1"
mccabe = "^0.6"
Pillow = "^8.3"
psycopg2 = "^2.9"
PyJWT = "^2.3"
pyrsistent = "^0.18"
python-decouple = "^3.4"
pytz = "^2021.1"
PyYAML = "^6.0"
sqlparse = "^0.4"
static3 = "^0.7"
typing-extensions = "^3.10.0"
uritemplate = "^4.1"
zipp = "^3.6"
requests = "^2.26.0"
vcrpy = "^4.1.1"

[tool.poetry.dev-dependencies]
pycodestyle = "^2.8"
pyflakes = "^2.4"
black = "^22.1.0"
isort = "^5.10.1"
pytest = "^7.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
